Updated the project to compile against Spring boot 2 and Spring framework 5.
Even though there are no changes in the main source files, it was neccessary to update the project since it needs to rely on an updated implementation of the Persistable interface (from Spring Data 2.0.8). Otherwise, a NoSuchMethodError is thrown when this library tries to use the getId() method of it.