Implemented auto-generation of BuilderConstructors
AbstractBuilder can now automatically generate the required
`BuilderConstructors` used to obtain references to a `BuildCommand` given an
Entity. This is done by accessing the no-args constructor of the entity
and the BuildCommand class.

For customization and backwards compatibility, it's still possible to
override the `constructors()` method of `AbstractBuilder` and supply
your own set of `BuilderConstructors`

Fixes #4