Now automatically retrieves a JpaRepository for the given entity class
When creating a BuildCommand, it is no longer required to override the
method `getRepository()`. The repository is now looked up using the
`Repositories` class from the Spring Framework.

If no repository is available, the builder still works. You can also
still override the `getRepository()` method in your buildCommand to
return a customized repository instance (or disable the repository on
purpose)

Fixes #6