Implemented auto-generation of `with` methods
The Heph library has been enhanced with the capability of generating
`setter` methods for your buildCommand. This is done by using a Spring AOP
proxy, based on the work of [Beanie](https://github.com/42BV/beanie).

This means you only have to declare the methods to apply values to your
entity in your BuildCommand implementation.

However, it also means a breaking change for existing builders. See
[CHANGELOG.MD](https://github.com/42BV/heph/blob/master/CHANGELOG.md)
for more info and a migration guide.

Also, the logging dependencies have been removed from `pom.xml`, as
these are already provided by the nested `spring-boot-starter-logging`
of `spring-boot-starter-data-jpa` and were flagged as duplicate.

Fixes #9