Fixed ClassCastException when passing a Supplier of a collection type
If an entity has a collection type of nested entities (e.g. `OneToMany`
or `ManyToMany` mapping), you should be able to pass the set of nested
entities as a supplied collection to prevent unneccessary entity
creation in deeper tree structures.

However, the type of `AbstractLazyEntity` was bound to Persistable,
preventing passing of a `Collection` inside it.

This commit removes that restriction and now allows passing Collections
in a Supplier function. A test case has been written as well.

Fixes #14