Removed hard dependency on BeanMapper
BeanMapper is only required for the `copy` method, but currently it was
a hard dependency of this library. To make things worse, its scope was
`provided`, so you would get a runtime crash if your project did not
include `BeanMapper` by itself.

This has now been solved. If BeanMapper is *not* on the classpath,
everything but the `copy` method can still be used.