Upgraded to Java 11
Java 8 is no longer supported in this version!

The previous way of reflection used in BuildCommandAdvice is no longer
working in Java 11. This has been replaced with a way that does work.

Setting Arrays also no longer worked. A field of type String[] could
no longer be filled with an Object[] as was possible in Java 8. This
bit has been improved to fill the field with a String[]. The code
detects of what type the field is and creates an Array of said type
accordingly, keeping this dynamic for all types of arrays.